move server res into lib/server and test/server
