+++ bin/server.js
@@ -10,9 +10,9 @@ const assert  = require('assert');
 const util    = require('util');
 const MongoClient = require('mongodb').MongoClient;
 const GetOpt      = require('node-getopt');
-const LOGGER      = require('../lib/logsetup.js');
+const LOGGER      = require('../lib/server/logsetup.js');
 
-const RangerController = require('../lib/controller');
+const RangerController = require('../lib/server/controller');
 
 var opt = new GetOpt([
     ['p','port=PORT'        ,'PORT or socket which server listens on'],
