+++ lib/server/auth.js
@@ -2,7 +2,7 @@
 
 /**
  * auth module.
- * @module lib/auth
+ * @module lib/server/auth
  *
  * @description Authorisation procedure.
  *              Relies on data access group being known for each data
@@ -10,7 +10,7 @@
  *              of data access groups.
  *
  * @example <caption>Example usage of the auth module.</caption>
- *   const DataAccess = require('../lib/auth.js');
+ *   const DataAccess = require('../lib/server/auth.js');
  *
  *   // Example of a database record mapping users to access groups
  *   var records_json = '[
