+++ lib/server/controller.js
@@ -2,12 +2,12 @@
 
 /**
  * controller module.
- * @module lib/controller
+ * @module lib/server/controller
  *
  * @description Controller module for the streaming HTTP server.
  *
  * @example <caption>Example usage of the controller.</caption>
- *   const RangerController = require('../lib/controller.js');
+ *   const RangerController = require('../lib/server/controller.js');
  *   // Create a new object, no authorisation.
  *   let c = new RangerController(request, response, db);
  *   // Create a new object with authorisation.
