+++ lib/server/logsetup.js
