+++ lib/server/mapper.js
@@ -2,7 +2,7 @@
 
 /**
  * mapper module.
- * @module lib/mapper
+ * @module lib/server/mapper
  *
  * @description Authorisation for access to data sources (sequencing files).
  *              Relies on data access group being known for each data
@@ -10,7 +10,7 @@
  *              of data access groups.
  *
  * @example <caption>Example usage of the mapper module.</caption>
- *   const DataAccess = require('../lib/mapper.js');
+ *   const DataAccess = require('../lib/server/mapper.js');
  *   // Create a new object.
  *   var da = new DataMapper(db);
  *   // Register listeners for events this object emits.
