+++ lib/server/model.js
@@ -2,12 +2,12 @@
 
 /**
  * model module.
- * @module lib/model
+ * @module lib/server/model
  *
  * @description Model for the streaming HTTP server.
  *
  * @example <caption>Example usage of the model.</caption>
- *   const RangerModel = require('../lib/model.js');
+ *   const RangerModel = require('../lib/server/model.js');
  *   // Create a new object.
  *   let m = new RangerModel();
  *   // Create a new object and define the temporary data directory.
@@ -26,7 +26,7 @@ const fse      = require('fs-extra');
 const os       = require('os');
 const path     = require('path');
 const LOGGER   = require('winston');
-const pipeline = require('../lib/pipeline.js');
+const pipeline = require('./pipeline.js');
 
 const SAMTOOLS_COMMAND     = 'samtools';
 const BBB_MARKDUPS_COMMAND = 'bamstreamingmarkduplicates';
