+++ lib/server/pipeline.js
@@ -2,7 +2,7 @@
 
 /**
  * pipeline module.
- * @module lib/pipeline
+ * @module lib/server/pipeline
  * @param {Array} processes an array of child processes
  * @param {Function} onSuccessClbk function to call if all processes succeed
  * @param {Function} onFailureClbk function to call if one of processes fails
@@ -14,7 +14,7 @@
  *              run asynchronously. Promises are used to get feedback.
  *
  * @example <caption>Example usage of the pipeline module.</caption>
- *   const pipeline = require('../lib/pipeline.js');
+ *   const pipeline = require('../lib/server/pipeline.js');
  *   const cat = spawn('cat', ['file']);
  *   const wc = spawn('wc', ['-l']);
  *   var success = function() {console.log('Success')};
