+++ test/server/data/fixtures/fileinfo.json
